Magnalium V0.03
Cave area implemented in the world which can be alternated with the forest area by clicking the world icon twice. Inventory system with amount of items stacking working. Recipe system with methods for drawing recipe UI's and setting up different items for each recipe. A Currency class for handling Currency in the game which can be used in the Store, which is a new system implemented where the player can buy new Recipes, the Store contains three tabs where the player can choose between new Axes, Pickaxes or Swords.